listboxex: gcc 4.7.2.

+++ listboxex/ListBoxEx.hpp
@@ -98,7 +98,7 @@
 #define LIFEX_DISABLE                   2
 #define LIFEX_HIDDEN                    4
 
-extern long WINAPI ListBoxExDialogProc(HANDLE hDlg,int Msg,int Param1,void* Param2);
+extern intptr_t WINAPI ListBoxExDialogProc(HANDLE hDlg,intptr_t Msg,intptr_t Param1,void* Param2);
 
 extern PluginStartupInfo Info;
 extern FARSTANDARDFUNCTIONS FSF;
@@ -138,7 +138,7 @@ struct ListBoxExData
 //structures for messages
 struct ListBoxExSetColor
 {
-  long Index;
+  intptr_t Index;
   unsigned long TypeIndex;
   unsigned long ColorIndex;
   ListBoxExColor Color;
@@ -146,31 +146,31 @@ struct ListBoxExSetColor
 
 struct ListBoxExSetHotkey
 {
-  long Index;
+  intptr_t Index;
   long Hotkey;
 };
 
 struct ListBoxExSetFlags
 {
-  long Index;
+  intptr_t Index;
   long Flags;
 };
 
 struct ListBoxExManageItem
 {
-  long Index;
+  intptr_t Index;
   ListBoxExItem Item;
 };
 
 struct ListBoxExAddItem
 {
-  long Count;
+  intptr_t Count;
   ListBoxExItem *Items;
 };
 
 struct ListBoxExColors
 {
-  long Count;
+  intptr_t Count;
   FarColor *Colors;
 };
